requirements.txt
mysql-connector-python
pymongo
redis
pyspark
kafka-python
python-dotenv
avro-python3
fastparquet
